requirements.txt
PyQt6==6.7.0
requests==2.31.0
urllib3<2.0
google-api-python-client==2.100.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
python-docx==1.1.0
python-dotenv==1.0.0
selenium==4.17.2
psutil>=5.9.0
beautifulsoup4>=4.12.0
slack-sdk>=3.27.0
